NSA backdoor - ignore

+++ src/main/java/com/kbi/qwertech/blocks/BlockCorrugated.java
@@ -72,7 +72,7 @@ public class BlockCorrugated extends BlockBaseMeta {
 	
 	
 	@Override
-	public float getExplosionResistance(int meta)
+	public float getExplosionResistance(byte meta)
 	{
 		return 3.3F * hardness[meta];
 	}
@@ -91,10 +91,7 @@ public class BlockCorrugated extends BlockBaseMeta {
 	
 	
 	@Override
-	public boolean isSealable(int aMeta, byte aSide)
-	{
-	   return false;
-	}
+	public boolean isSealable(byte aMeta, byte aSide) { return false; }
 	  
 	@Override
 	public boolean isFireSource(World aWorld, int aX, int aY, int aZ, ForgeDirection aSide)
